BiometryTraining: Functions to Assist Design and Analysis of Agronomic Experiments

Provides functions to aid in the design and analysis of agronomic and agricultural experiments through easy access to documentation and helper functions, especially for users who are learning these concepts.

See also

Author

Maintainer: Sam Rogers biometrytraining@adelaide.edu.au

Authors:

Sharon Nielsen biometrytraining@adelaide.edu.au

Annie Conway biometrytraining@adelaide.edu.au

Other contributors:

University of Adelaide [copyright holder, funder]

GRDC [copyright holder, funder]