Produces graph of design layout

Produces graph of design layout

# S3 method for des
plot(
  design.obj,
  nrows,
  ncols,
  brows,
  bcols,
  rotation,
  size,
  margin,
  return.seed
)

Arguments

design.obj
An agricolae design object.

nrows
The number of rows in the design.

ncols
The number of columns in the design.

brows
For RCBD only. The number of rows in a block.

bcols
For RCBD only. The number of columns in a block.

rotation
Rotate the text output as Treatments within the plot. Allows for easier reading of long treatment labels. Number between 0 and 360 (inclusive) - default 0

size
Increase or decrease the text size within the plot for treatment labels. Numeric with default value of 4.

margin
Logical (default FALSE). Expand the plot to the edges of the plotting area i.e. remove whitespace between plot and axes.

return.seed
Logical (default TRUE). Output the seed used in the design?

quiet
Logical (default FALSE). Return the objects without printing output.

Value

Returns dataframe of design and ggplot object of design layout.