Tukey's Honest Significant Difference test

Performs Tukey's Honest Significant Difference test and optionally back-transforms natural log and squareroot transformed predicted values.

tuk.out(pred.obj, model.obj, pred, sig = 0.95, trans = FALSE, offset = 0)

Arguments

pred.obj
An ASReml-R prediction object.

model.obj
An ASReml-R model object.

pred
A character string of the main effect or interaction terms predicted from the model.

sig
The significance level to test at.

trans
Choice of log or sqrt to back-transform predicted values and standard errors in the output.

offset
A numeric value used in the transformation.

Value

A data frame with the predicted values and results of Tukey's test.

References

http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.47.9023

Examples

if (FALSE) dat.asr <- asreml(Yield ~ Variety, random = ~ Block,
residual = ~ id(Plot), data = dat)

dat.pred <- predict(dat.asr, classify = "Variety",
                    sed = TRUE)

#> Error in predict(dat.asr, classify = "Variety", sed = TRUE): object 'dat.asr' not found

pred.out <- tuk.out(model.obj = dat.asr, pred.obj = dat.pred,
                   data = dat, pred = "Variety", sig = 0.95)

#> Error in tuk.out(model.obj = dat.asr, pred.obj = dat.pred, data = dat,     pred = "Variety", sig = 0.95): unused argument (data = dat)

pred.out

#> Error in eval(expr, envir, enclos): object 'pred.out' not found